+++ walkers/src/lib.rs
@@ -7,12 +7,14 @@ pub mod extras;
 mod io;
 mod map;
 mod mercator;
+mod position;
 pub mod sources;
 mod tiles;
 mod zoom;
 
 pub use download::{HeaderValue, HttpOptions};
 pub use map::{Map, MapMemory, Plugin, Projector};
-pub use mercator::{lat_lon, lon_lat, screen_to_position, Position, TileId};
+pub use mercator::{screen_to_position, TileId};
+pub use position::{lat_lon, lon_lat, Position};
 pub use tiles::{HttpTiles, Texture, TextureWithUv, Tiles};
 pub use zoom::InvalidZoom;
